requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.24.3
pandas==2.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
river==0.16.0
scipy==1.10.1
six==1.16.0
tzdata==2023.3
Werkzeug==1.0.1
